chore: remove white-box tests
